+++ packages/telescope-posts/lib/client/templates/modules/post_admin.js
@@ -1,9 +1,9 @@
 Template.post_admin.helpers({
   showApprove: function () {
-    return this.status === Posts.config.STATUS_PENDING;
+    return !!Settings.get('requirePostsApproval') && (this.status === Posts.config.STATUS_PENDING || this.status === Posts.config.STATUS_REJECTED);
   },
-  showUnapprove: function(){
-    return !!Settings.get('requirePostsApproval') && this.status === Posts.config.STATUS_APPROVED;
+  showReject: function(){
+    return !!Settings.get('requirePostsApproval') && (this.status === Posts.config.STATUS_PENDING || this.status === Posts.config.STATUS_APPROVED);
   },
   shortScore: function(){
     return Math.floor(this.score*100)/100;
@@ -15,8 +15,8 @@ Template.post_admin.events({
     Meteor.call('approvePost', this._id);
     e.preventDefault();
   },
-  'click .unapprove-link': function(e){
-    Meteor.call('unapprovePost', this._id);
+  'click .reject-link': function(e){
+    Meteor.call('rejectPost', this._id);
     e.preventDefault();
   },
   'click .delete-link': function(e){