+++ packages/telescope-posts/lib/modules.js
@@ -2,7 +2,7 @@
 Telescope.modules.add("top", {
   template: 'views_menu',
   order: 99,
-  only: ["posts_default", "posts_top", "posts_new", "posts_best", "posts_pending", "posts_scheduled"]
+  only: ["posts_default", "posts_top", "posts_new", "posts_best", "posts_pending", "posts_rejected", "posts_scheduled"]
 });
 
 Telescope.modules.add("postComponents", [