+++ packages/telescope-posts/lib/routes.js
@@ -95,6 +95,13 @@ Posts.controllers.pending = Posts.controllers.list.extend({
 });
 
 /**
+ * Controller for rejected view
+ */
+Posts.controllers.rejected = Posts.controllers.list.extend({
+  view: 'rejected'
+});
+
+/**
  * Controller for scheduled view
  */
 Posts.controllers.scheduled = Posts.controllers.list.extend({
@@ -229,6 +236,13 @@ Meteor.startup(function () {
   Router.route('/pending/:limit?', {
     name: 'posts_pending',
     controller: Posts.controllers.pending
+  });
+
+  // Rejected
+
+  Router.route('/rejected/:limit?', {
+    name: 'posts_rejected',
+    controller: Posts.controllers.rejected
   });
 
   // Scheduled